Move parallels executor integration tests to own file
